+++ node/src/test/test-Router.ts
@@ -2,11 +2,12 @@ import * as mediasoup from '../';
 import { enhancedOnce } from '../enhancedEvents';
 import type { WorkerImpl } from '../Worker';
 import type { WorkerEvents, RouterEvents } from '../types';
-import { InvalidStateError } from '../errors';
+import { InvalidStateError, UnsupportedError } from '../errors';
 import * as utils from '../utils';
 
 type TestContext = {
 	mediaCodecs: mediasoup.types.RtpCodecCapability[];
+	unsupportedMediaCodecs: mediasoup.types.RtpCodecCapability[];
 	worker?: mediasoup.types.Worker;
 };
 
@@ -39,6 +40,21 @@ const ctx: TestContext = {
 			rtcpFeedback: [], // Will be ignored.
 		},
 	]),
+	unsupportedMediaCodecs: utils.deepFreeze<
+		mediasoup.types.RtpCodecCapability[]
+	>([
+		{
+			kind: 'audio',
+			mimeType: 'audio/x-aiff',
+			clockRate: 8000,
+			channels: 1,
+		},
+		{
+			kind: 'video',
+			mimeType: 'video/3gpp',
+			clockRate: 90000,
+		},
+	]),
 };
 
 beforeEach(async () => {
@@ -69,6 +85,8 @@ test('worker.createRouter() succeeds', async () => {
 	expect(router.closed).toBe(false);
 	expect(typeof router.rtpCapabilities).toBe('object');
 	expect(Array.isArray(router.rtpCapabilities.codecs)).toBe(true);
+	// 3 codecs + 2 RTX codecs.
+	expect(router.rtpCapabilities.codecs?.length).toBe(5);
 	expect(Array.isArray(router.rtpCapabilities.headerExtensions)).toBe(true);
 	expect(router.appData).toEqual({ foo: 123 });
 
@@ -106,6 +124,12 @@ test('worker.createRouter() succeeds', async () => {
 	expect((ctx.worker! as WorkerImpl).routersForTesting.size).toBe(0);
 }, 2000);
 
+test('worker.createRouter() with invalid codecs rejects with UnsupportedError', async () => {
+	await expect(
+		ctx.worker!.createRouter({ mediaCodecs: ctx.unsupportedMediaCodecs })
+	).rejects.toThrow(UnsupportedError);
+}, 2000);
+
 test('worker.createRouter() with wrong arguments rejects with TypeError', async () => {
 	// @ts-expect-error --- Testing purposes.
 	await expect(ctx.worker!.createRouter({ mediaCodecs: {} })).rejects.toThrow(
@@ -124,6 +148,25 @@ test('worker.createRouter() rejects with InvalidStateError if Worker is closed',
 	await expect(
 		ctx.worker!.createRouter({ mediaCodecs: ctx.mediaCodecs })
 	).rejects.toThrow(InvalidStateError);
+}, 2000);
+
+test('router.updateMediaCodecs() succeeds', async () => {
+	const router = await ctx.worker!.createRouter({
+		mediaCodecs: ctx.mediaCodecs,
+	});
+
+	expect(typeof router.rtpCapabilities).toBe('object');
+	expect(Array.isArray(router.rtpCapabilities.codecs)).toBe(true);
+	// 3 codecs + 2 RTX codecs.
+	expect(router.rtpCapabilities.codecs?.length).toBe(5);
+	expect(Array.isArray(router.rtpCapabilities.headerExtensions)).toBe(true);
+
+	router.updateMediaCodecs([]);
+
+	expect(typeof router.rtpCapabilities).toBe('object');
+	expect(Array.isArray(router.rtpCapabilities.codecs)).toBe(true);
+	expect(router.rtpCapabilities.codecs?.length).toBe(0);
+	expect(Array.isArray(router.rtpCapabilities.headerExtensions)).toBe(true);
 }, 2000);
 
 test('router.close() succeeds', async () => {