+++ rust/examples/echo.rs
@@ -237,7 +237,7 @@ impl Actor for EchoConnection {
                 ice_candidates: self.transports.producer.ice_candidates().clone(),
                 ice_parameters: self.transports.producer.ice_parameters().clone(),
             },
-            router_rtp_capabilities: self.router.rtp_capabilities().clone(),
+            router_rtp_capabilities: (*self.router.rtp_capabilities().read()).clone(),
         };
 
         ctx.address().do_send(server_init_message);